refactor : firebase 관련 함수 firebaseController에 모으기
