+++ src/utils/firebaseController.ts
@@ -1,5 +1,5 @@
-import { FIREBASE_CONFIG } from '@/constants';
 import { initializeApp } from 'firebase/app';
+import { getAuth } from 'firebase/auth';
 import {
   collection,
   deleteDoc,
@@ -10,13 +10,22 @@ import {
   setDoc,
 } from 'firebase/firestore';
 
-const getDB = () => {
-  const firebaseDB = initializeApp(FIREBASE_CONFIG);
-  return getFirestore(firebaseDB);
+const FIREBASE_CONFIG = {
+  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
+  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
+  databaseURL: process.env.NEXT_PUBLIC_FIREBASE_DATABASE_URL,
+  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
+  measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID,
 };
 
+const app = initializeApp(FIREBASE_CONFIG);
+const db = getFirestore(app);
+const auth = getAuth(app);
+
 export const getData = async (table: string) => {
-  const db = getDB();
   const q = query(collection(db, table));
   const querySnapshot = await getDocs(q);
 
@@ -26,7 +35,6 @@ export const getData = async (table: string) => {
 };
 
 export const postData = async (table: string, data: any) => {
-  const db = getDB();
   const ref = collection(db, table);
 
   if (data.id) await setDoc(doc(ref, data.id), data);
@@ -37,12 +45,10 @@ export const postData = async (table: string, data: any) => {
 };
 
 export const deleteData = async (table: string, key: string) => {
-  const db = getDB();
   await deleteDoc(doc(db, table, key));
 };
 
 export const patchData = async (table: string, key: string, data: any) => {
-  const db = getDB();
   const ref = collection(db, table);
   await setDoc(doc(ref, key), data);
 };
